requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
google-genai>=1.0.0
pypdf>=5.0.0
slowapi>=0.1.9
bleach>=6.1.0
email-validator>=2.2.0
